+++ components/02-elements/dropdown-list/dropdown-list.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   status: 'wip',
   context: {
-    script: true,
+    id: 'dropdown-default',
     dropdowns: [
       {
         itemTag: 'a',